Relocates ConvertSpaceName method from network to core/network package. Removes needless regex export for invalid space chars.
